flat "log.level" field; use name for "log.logger" field (#23)
Per https://github.com/elastic/ecs-logging/blob/c3567a9/spec/spec.json#L21 it
is preferred to have "log.level" as a flat top-level field.

Also, the intent of the "log.logger" field is to be the app's logger name, rather
than the name of the logging framework. Pino supports a somewhat special "name"
field, which we'll translate to "log.logger". Winston and Morgan do not have a
core concept of a logger log, so we'll drop setting "log.logger" for them.

This change marks @elastic/ecs-pino-format@0.2.0, 
@elastic/ecs-winston-format@0.3.0, and @elastic/ecs-morgan-format@0.3.0.