tests(helpers): fix hang in test from having dropped usage of simple-get
In node v14 and later this test would hang waiting for the 'end' event
on the client request. This happened because the server was setting
Content-Length:42, but only sending 2 bytes ("ok"). Before the recent
change dropping the "simple-get" module this passed because "simple-get"
was overwriting the Content-Length header with the actual length of the
given post body.