tests: ensure we are testing field order ("index") in ecs-logging validation
ecsLoggingValidate can only validate the "index" spec field if given a
log record *string*. This changes the tests to pass in a string rather
than a preparsed object. This brought up that pino does not satisfy the
"index" -- a known issue -- so support for explicitly skipping that
part of validation was added.