tests: fix APM tests for elastic-apm-node@3.16.0 (#83)
The APM tests for the Winston and Pino formatters started failing with
the release of elastic-apm-node@3.16.0, because it included changes that
could lead to the order of APM reported "transaction" and "span" objects
being reverse (in this case because span encoding is now faster). This
changes the test code to not be subject to that order race.