tests: fix apm.test.js breakage with elastic-apm-node@3.22.0 (#106)
In elastic/apm-agent-nodejs#2294 improve the behaviour of apm.flush()
to wait for inflight spans. This improvement broke a faulty assumption
in the apm.test.js code for pino and winston: those tests assumed that
calling `apm.flush()` as they did would result in a single intake
request with both transaction and span. What actually happened with the
newer flush implementation is *two* intake requests: one with the span
(explicitly flushed by the apm.flush() call), and the second with the
transaction (flushed on process exit after the request completed).