chore(slack-lite): switch to pull_request_target so have access to repo secrets (#208)
No need for us to be running code from the PR itself, so this event is fine.
Before this change, a workflow run for a PR from a *fork* would have
'Secret source: None' and hence not have access to the SLACK_BOT_TOKEN

Also be clearer on the 'slack-lite' brand.

Refs: https://github.com/elastic/apm-agent-nodejs/pull/4501